requirements.txt
python-telegram-bot==20.1
langchain
langchain-community
python-dotenv
apscheduler
openai